Add task 3 hot fix

+++ 0x1A-hash_tables/2-key_index.c
@@ -8,7 +8,7 @@
  */
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	unsigned int hash, index;
+	unsigned long int hash, index;
 
 	hash = hash_djb2(key);
 	index = hash % size;
